Refactor StackedPointsChart into class component
Refactor StackedPointsChart into class component and add data handling from PlayerOverview component.
